All tailwind settings needed fore the component are now in a plugin
The purpose of that plugin is to make sure the component always work. It will doing so by transferring initials the necessaries into a tailwind plugin that will be necessary in the `tailwind.config.js` file of each projects that will depend on the component repo.

It will be possible to