Added viewport to inline icons
For some reasons, adding any properties to imported svg using `react-svg-loader` remove the property `viewBox`. This is what caused the searchbox to be broken… For the moment, the best fix is to simply add it.

The best solution would be to use `react-inlinesvg` everywhere, but it has issue with `webpack` export.

I suggest to leave it as is for the moment